Inject auto-generated polling places in elections for test decks
Factor out the common logic for both exporting the election package and test decks.
